requirements.txt
SpeechRecognition
openai
python-dotenv
pyaudio
pyttsx3